requirements.txt
python-dotenv
pandas
numpy<2.0
pymongo
certifi
pymongo[srv]
scikit-learn
dill
mlflow
dagshub
dvc
fastapi
uvicorn
python-multipart
geopy
folium
plotly
streamlit

## -e .